+++ Bugsnag/Bugsnag.m
@@ -62,6 +62,10 @@ BSG_OBJC_DIRECT_MEMBERS
     }
 }
 
++ (BOOL)isStarted {
+    return bsg_g_bugsnag_client.isStarted;
+}
+
 /**
  * Purge the global client so that it will be regenerated on the next call to start.
  * This is only used by the unit tests.
@@ -75,51 +79,51 @@ BSG_OBJC_DIRECT_MEMBERS
 }
 
 + (BOOL)appDidCrashLastLaunch {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return [self.client appDidCrashLastLaunch];
     }
     return NO;
 }
 
 + (BugsnagLastRunInfo *)lastRunInfo {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return self.client.lastRunInfo;
     }
     return nil;
 }
 
 + (void)markLaunchCompleted {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client markLaunchCompleted];
     }
 }
 
 + (void)notify:(NSException *)exception {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client notify:exception];
     }
 }
 
 + (void)notify:(NSException *)exception block:(BugsnagOnErrorBlock)block {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client notify:exception block:block];
     }
 }
 
 + (void)notifyError:(NSError *)error {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client notifyError:error];
     }
 }
 
 + (void)notifyError:(NSError *)error block:(BugsnagOnErrorBlock)block {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client notifyError:error block:block];
     }
 }
 
-+ (BOOL)bugsnagStarted {
-    if (!self.client.started) {
++ (BOOL)bugsnagReadyForInternalCalls {
+    if (!self.client.readyForInternalCalls) {
         bsg_log_err(@"Ensure you have started Bugsnag with startWithApiKey: "
                     @"before calling any other Bugsnag functions.");
 
@@ -129,14 +133,14 @@ BSG_OBJC_DIRECT_MEMBERS
 }
 
 + (void)leaveBreadcrumbWithMessage:(NSString *)message {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client leaveBreadcrumbWithMessage:message];
     }
 }
 
 + (void)leaveBreadcrumbForNotificationName:
     (NSString *_Nonnull)notificationName {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client leaveBreadcrumbForNotificationName:notificationName];
     }
 }
@@ -145,7 +149,7 @@ BSG_OBJC_DIRECT_MEMBERS
                           metadata:(NSDictionary *_Nullable)metadata
                            andType:(BSGBreadcrumbType)type
 {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client leaveBreadcrumbWithMessage:message
                                        metadata:metadata
                                         andType:type];
@@ -154,13 +158,13 @@ BSG_OBJC_DIRECT_MEMBERS
 
 + (void)leaveNetworkRequestBreadcrumbForTask:(NSURLSessionTask *)task
                                      metrics:(NSURLSessionTaskMetrics *)metrics {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client leaveNetworkRequestBreadcrumbForTask:task metrics:metrics];
     }
 }
 
 + (NSArray<BugsnagBreadcrumb *> *_Nonnull)breadcrumbs {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return self.client.breadcrumbs;
     } else {
         return @[];
@@ -168,19 +172,19 @@ BSG_OBJC_DIRECT_MEMBERS
 }
 
 + (void)startSession {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client startSession];
     }
 }
 
 + (void)pauseSession {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client pauseSession];
     }
 }
 
 + (BOOL)resumeSession {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return [self.client resumeSession];
     } else {
         return false;
@@ -192,31 +196,31 @@ BSG_OBJC_DIRECT_MEMBERS
 // =============================================================================
 
 + (void)addFeatureFlagWithName:(NSString *)name variant:(nullable NSString *)variant {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client addFeatureFlagWithName:name variant:variant];
     }
 }
 
 + (void)addFeatureFlagWithName:(NSString *)name {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client addFeatureFlagWithName:name];
     }
 }
 
 + (void)addFeatureFlags:(NSArray<BugsnagFeatureFlag *> *)featureFlags {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client addFeatureFlags:featureFlags];
     }
 }
 
 + (void)clearFeatureFlagWithName:(NSString *)name {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client clearFeatureFlagWithName:name];
     }
 }
 
 + (void)clearFeatureFlags {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client clearFeatureFlags];
     }
 }
@@ -237,7 +241,7 @@ BSG_OBJC_DIRECT_MEMBERS
             withKey:(NSString *_Nonnull)key
           toSection:(NSString *_Nonnull)section
 {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client addMetadata:metadata
                                   withKey:key
                                 toSection:section];
@@ -247,7 +251,7 @@ BSG_OBJC_DIRECT_MEMBERS
 + (void)addMetadata:(id _Nonnull)metadata
           toSection:(NSString *_Nonnull)section
 {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client addMetadata:metadata
                        toSection:section];
     }
@@ -255,7 +259,7 @@ BSG_OBJC_DIRECT_MEMBERS
 
 + (NSMutableDictionary *)getMetadataFromSection:(NSString *)section
 {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return [[self.client getMetadataFromSection:section] mutableCopy];
     }
     return nil;
@@ -264,7 +268,7 @@ BSG_OBJC_DIRECT_MEMBERS
 + (id _Nullable )getMetadataFromSection:(NSString *_Nonnull)section
                                 withKey:(NSString *_Nonnull)key
 {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return [[self.client getMetadataFromSection:section withKey:key] mutableCopy];
     }
     return nil;
@@ -272,7 +276,7 @@ BSG_OBJC_DIRECT_MEMBERS
 
 + (void)clearMetadataFromSection:(NSString *)section
 {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client clearMetadataFromSection:section];
     }
 }
@@ -280,7 +284,7 @@ BSG_OBJC_DIRECT_MEMBERS
 + (void)clearMetadataFromSection:(NSString *_Nonnull)sectionName
                          withKey:(NSString *_Nonnull)key
 {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client clearMetadataFromSection:sectionName
                                       withKey:key];
     }
@@ -289,13 +293,13 @@ BSG_OBJC_DIRECT_MEMBERS
 // MARK: -
 
 + (void)setContext:(NSString *_Nullable)context {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client setContext:context];
     }
 }
 
 + (NSString *_Nullable)context {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return self.client.context;
     }
     return nil;
@@ -308,13 +312,13 @@ BSG_OBJC_DIRECT_MEMBERS
 + (void)setUser:(NSString *_Nullable)userId
       withEmail:(NSString *_Nullable)email
         andName:(NSString *_Nullable)name {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client setUser:userId withEmail:email andName:name];
     }
 }
 
 + (nonnull BugsnagOnSessionRef)addOnSessionBlock:(nonnull BugsnagOnSessionBlock)block {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return [self.client addOnSessionBlock:block];
     } else {
         // We need to return something from this nonnull method; simulate what would have happened.
@@ -323,14 +327,14 @@ BSG_OBJC_DIRECT_MEMBERS
 }
 
 + (void)removeOnSession:(nonnull BugsnagOnSessionRef)callback {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client removeOnSession:callback];
     }
 }
 
 + (void)removeOnSessionBlock:(BugsnagOnSessionBlock _Nonnull )block
 {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client removeOnSessionBlock:block];
     }
 }
@@ -340,7 +344,7 @@ BSG_OBJC_DIRECT_MEMBERS
 // =============================================================================
 
 + (nonnull BugsnagOnBreadcrumbRef)addOnBreadcrumbBlock:(nonnull BugsnagOnBreadcrumbBlock)block {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         return [self.client addOnBreadcrumbBlock:block];
     } else {
         // We need to return something from this nonnull method; simulate what would have happened.
@@ -349,13 +353,13 @@ BSG_OBJC_DIRECT_MEMBERS
 }
 
 + (void)removeOnBreadcrumb:(nonnull BugsnagOnBreadcrumbRef)callback {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client removeOnBreadcrumb:callback];
     }
 }
 
 + (void)removeOnBreadcrumbBlock:(BugsnagOnBreadcrumbBlock _Nonnull)block {
-    if ([self bugsnagStarted]) {
+    if ([self bugsnagReadyForInternalCalls]) {
         [self.client removeOnBreadcrumbBlock:block];
     }
 }