+++ features/fixtures/shared/scenarios/Scenario.h
@@ -15,12 +15,12 @@ void markErrorHandledCallback(const BSG_KSCrashReportWriter *writer);
 
 @interface Scenario : NSObject
 
-@property (class, readonly) NSURL *mazeRunnerURL;  
-
 @property (strong, nonatomic, nonnull) BugsnagConfiguration *config;
 
-+ (Scenario *)createScenarioNamed:(NSString *)className withConfig:(nullable BugsnagConfiguration *)config;
-
++ (Scenario *)createScenarioNamed:(NSString *)className
+                       withConfig:(nullable BugsnagConfiguration *)config;
+    
+@property (class, readwrite) NSString *baseMazeAddress;
 @property (class, readonly, nullable) Scenario *currentScenario;
 
 - (instancetype)initWithConfig:(nullable BugsnagConfiguration *)config;