+++ features/fixtures/shared/scenarios/Scenario.m
@@ -7,14 +7,11 @@
 #import <objc/runtime.h>
 
 #if TARGET_OS_IOS
-#define MAZE_RUNNER_URL "http://bs-local.com:9339"
 #define SWIFT_MODULE "iOSTestApp"
 #elif TARGET_OS_OSX
-#define MAZE_RUNNER_URL "http://localhost:9339"
 #define SWIFT_MODULE "macOSTestApp"
 #elif TARGET_OS_WATCH
 #import "watchos_maze_host.h"
-#define MAZE_RUNNER_URL "http://" WATCHOS_MAZE_HOST ":9339"
 #define SWIFT_MODULE "watchOSTestApp_WatchKit_Extension"
 #else
 #error Unsupported TARGET_OS
@@ -35,6 +32,7 @@ void markErrorHandledCallback(const BSG_KSCrashReportWriter *writer) {
 // MARK: -
 
 static Scenario *theScenario;
+static NSString *theBaseMazeAddress;
 
 #if !TARGET_OS_WATCH
 static char ksLogPath[PATH_MAX];
@@ -71,10 +69,6 @@ static char ksLogPath[PATH_MAX];
     }];
 }
 
-+ (NSURL *)mazeRunnerURL {
-    return [NSURL URLWithString:@MAZE_RUNNER_URL];
-}
-
 + (Scenario *)createScenarioNamed:(NSString *)className withConfig:(BugsnagConfiguration *)config {
     Class class = NSClassFromString(className) ?:
     NSClassFromString([@SWIFT_MODULE "." stringByAppendingString:className]);
@@ -90,14 +84,22 @@ static char ksLogPath[PATH_MAX];
     return theScenario;
 }
 
++ (NSString *)baseMazeAddress {
+    return theBaseMazeAddress;
+}
+
++ (void)setBaseMazeAddress:(NSString *)baseMazeAddress {
+    theBaseMazeAddress = baseMazeAddress;
+}
+
 - (instancetype)initWithConfig:(BugsnagConfiguration *)config {
     if (self = [super init]) {
         if (config) {
             _config = config;
         } else {
             _config = [[BugsnagConfiguration alloc] initWithApiKey:@"12312312312312312312312312312312"];
-            _config.endpoints.notify = @MAZE_RUNNER_URL "/notify";
-            _config.endpoints.sessions = @MAZE_RUNNER_URL "/sessions";
+            _config.endpoints.notify = [NSString stringWithFormat:@"%@/notify", theBaseMazeAddress];
+            _config.endpoints.sessions = [NSString stringWithFormat:@"%@/sessions", theBaseMazeAddress];
         }
 #if !TARGET_OS_WATCH
         _config.enabledErrorTypes.ooms = NO;
@@ -218,8 +220,8 @@ static NSURLSessionUploadTask * uploadTaskWithRequest_fromData_completionHandler
 + (void)executeMazeRunnerCommand:(void (^)(NSString *action, NSString *scenarioName, NSString *scenarioMode))preHandler {
     NSLog(@"%s", __PRETTY_FUNCTION__);
     NSURLSession *session = [NSURLSession sessionWithConfiguration:[NSURLSessionConfiguration ephemeralSessionConfiguration]];
-    
-    NSURLRequest *request = [NSURLRequest requestWithURL:[NSURL URLWithString:@MAZE_RUNNER_URL "/command"]];
+    NSString *commandEndpoint = [NSString stringWithFormat:@"%@/command", theBaseMazeAddress];
+    NSURLRequest *request = [NSURLRequest requestWithURL:[NSURL URLWithString:commandEndpoint]];
     [[session dataTaskWithRequest:request completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
         if (![response isKindOfClass:[NSHTTPURLResponse class]] || [(NSHTTPURLResponse *)response statusCode] != 200) {
             NSLog(@"%s request failed with %@", __PRETTY_FUNCTION__, response ?: error);
@@ -264,7 +266,6 @@ static NSURLSessionUploadTask * uploadTaskWithRequest_fromData_completionHandler
     
     [self startBugsnagForScenario:scenarioName eventMode:eventMode];
     
-    NSLog(@"Running scenario \"%@\"", NSStringFromClass([theScenario class]));
     [theScenario run];
 }
 
@@ -274,7 +275,7 @@ static NSURLSessionUploadTask * uploadTaskWithRequest_fromData_completionHandler
     theScenario = [Scenario createScenarioNamed:scenarioName withConfig:nil];
     theScenario.eventMode = eventMode;
     
-    NSLog(@"Starting scenario \"%@\"", NSStringFromClass([theScenario class]));
+    NSLog(@"Starting scenario \"%@\"", NSStringFromClass([self class]));
     [theScenario startBugsnag];
 }