+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 Changelog
 =========
 
+## 6.26.0 (2023-03-08)
+
+### Enhancements
+
+* Added isStarted to Bugsnag and BugsnagClient
+  [1528](https://github.com/bugsnag/bugsnag-cocoa/pull/1528)
+
+### Bug fixes
+
+* Fixed some race conditions that could cause issues in rare cases.
+  [1529](https://github.com/bugsnag/bugsnag-cocoa/pull/1529)
+
+* onError blocked from running if releaseStage not in enabledReleaseStages.
+  [1518](https://github.com/bugsnag/bugsnag-cocoa/pull/1518)
+
+
 ## 6.25.2 (2023-01-18)
 
 ### Bug fixes